+++ backends/npu/tests/unittests/test_gather_op_npu.py
@@ -85,8 +85,8 @@ class API_TestGather(unittest.TestCase):
             out = paddle.gather(data1, index)
             place = paddle.CustomPlace("npu", 0)
             exe = fluid.Executor(place)
-            input = np.array([[1, 2], [3, 4], [5, 6]])
-            index_1 = np.array([1, 2])
+            input = np.array([[1, 2], [3, 4], [5, 6]]).astype("float32")
+            index_1 = np.array([1, 2]).astype("int32")
             (result,) = exe.run(
                 feed={"data1": input, "index": index_1}, fetch_list=[out]
             )
